refactor: fn_traits: replace InplaceFnOnce with dummy allocator approach
Replace the unsafe InplaceFnOnce trait with a safer implementation that
uses a dummy allocator to create temporary Box instances. This leverages
existing Box<dyn FnOnce> implementations instead of reimplementing
complex dynamic dispatch logic.

Key changes:
- Add impl_fn_traits.rs with FnOnce/FnMut/Fn implementations
- Remove InplaceFnOnce trait and its unsafe call_once_impl method
- Update tests to use standard FnOnce trait instead of InplaceFnOnce
- Use allocator_api feature for Box::from_raw_in functionality

---

- Fix: #2
